improved ui, fixed CheckDialogFragment bugs, improved saved sorts

+++ app/src/main/java/ru/savenkov/paychecksapp/converter/Converter.kt
@@ -149,7 +149,7 @@ object Converter {
             item.shiftNumber,
             item.fiscalDocumentNumber,
             item.fiscalDriveNumber,
-            item.fiscalSign.toInt(),
+            item.fiscalSign,
             item.kktRegId,
             item.numberKkt,
             item.region,
